fix(engine): saturating arithmetic, comment cleanup, and miscellaneous hardening
Fourteen files with targeted safety and quality improvements:

- stdlib: Math.round uses saturating_add; Array.splice no-args edge
  case; opt_str_arg treats Undefined as None
- bytecode_vm: ObjectId uses u32::try_from().unwrap_or(u32::MAX)
- tropical_semiring: widen to i128 to prevent overflow
- performance_statistical_validation: .min(u32::MAX as u64) before
  truncation
- third_party_verifier: Display impl for VerificationVerdict
- parser_api_stability: unwrap_or_default() to expect() on serialize
- anti_entropy, bulkhead, hardware_code_layout_gate, shipped_path_matrix,
  law_proof_refutation, artifact_compression_pipeline: replace "in a
  real system" hedging comments with actionable "future work" notes
- react_compile_operator_surface: doc improvements, new Deprecated test
- jsx_tsx_parser: guard cursor.advance() with if let Some
- parser: comment cleanup in test
- ir_contract: expanded serde roundtrip test coverage for Ir3Instruction

+++ crates/franken-engine/src/tropical_semiring.rs
@@ -810,7 +810,7 @@ impl DeadCodeEliminator {
         }
 
         let elimination_ratio_millionths = if total_nodes > 0 {
-            (dead_indices.len() as i64 * 1_000_000) / total_nodes as i64
+            (dead_indices.len() as i128 * 1_000_000 / total_nodes as i128) as i64
         } else {
             0
         };
